requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.6.4
tortoise-orm==0.20.0
asyncpg==0.29.0
aerich==0.7.2
pytz==2024.1